LPS-JASPER Temp
